adding media files
I'd like to change the home page hero using the video background, the ratio is 16:9, we can do background-size: cover, and center, center.

also adding the SEO OG img for when someone going to tweet about the ceremony.

Unable to run the local at this moment, will come back to help.